test: remove node: imports from browser-run test files
Real browser mode runs test code inside the browser, unlike happy-dom
(which vitest executes inside Node), so any node: import in a
browser-run file breaks under real browsers.

- Move the 3 tests in mediaLibrary.test.ts that read real fixture
  files via node:fs into a new mediaLibrary.node.test.ts.
- Move fixture() and its node:path import out of client/helpers.ts
  into a new Node-only client/helpers.node.ts, imported only by
  *.node.test.ts files. helpers.ts now has zero node: imports.
- Rename distGraph.test.ts to distGraph.node.test.ts to keep it out
  of browser runs. A later commit relocates it into the packaging
  suite, where it belongs.
